move semantic kernel plugin prompt target out of pyrit and into the doc folder
