+++ doc/demo/xpia_helpers.py
