+++ pyrit/prompt_target/__init__.py
@@ -8,7 +8,6 @@ from pyrit.prompt_target.prompt_chat_target.azure_ml_chat_target import AzureMLC
 from pyrit.prompt_target.prompt_chat_target.openai_chat_target import AzureOpenAIChatTarget, OpenAIChatTarget
 from pyrit.prompt_target.gandalf_target import GandalfTarget
 from pyrit.prompt_target.text_target import TextTarget
-from pyrit.prompt_target.semantic_kernel_plugin_prompt_target import SemanticKernelPluginAzureOpenAIPromptTarget
 from pyrit.prompt_target.image_target import ImageTarget
 
 
@@ -22,5 +21,4 @@ __all__ = [
     "PromptChatTarget",
     "PromptTarget",
     "TextTarget",
-    "SemanticKernelPluginAzureOpenAIPromptTarget",
 ]
